rename to dependency-metrics
