+++ dependency_metrics/constants.py
