+++ dependency_metrics/package_managers/pip.py
