+++ pyproject.toml
@@ -3,7 +3,7 @@ build-backend = "setuptools.build_meta"
 requires = ["setuptools"]
 
 [project]
-name = "package-metrics"
+name = "dependency-metrics"
 version = "0.1.0"
 readme = "README.md"
 description = "Python tool to track outdated dependencies."
@@ -27,7 +27,7 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 [project.scripts]
-package-metrics = "package_metrics.metrics:main"
+dependency-metrics = "dependency_metrics.metrics:main"
 
 [project.optional-dependencies]
 test = [
@@ -36,4 +36,4 @@ test = [
 ]
 
 [project.urls]
-repository = "https://github.com/dimagi/package-metrics"
+repository = "https://github.com/dimagi/dependency-metrics"
